Split off related groups of endpoints into separate files
This also changes how we export the query/mutation hooks to be more robust to endpoints moving around. The code also reads a bit nicer after this change too.
