+++ src/app/components/navigation/IsaacApp.tsx
@@ -2,13 +2,13 @@ import React, {lazy, Suspense, useEffect} from 'react';
 import {
     AppState,
     fetchGlossaryTerms,
-    isaacApi,
     openActiveModal,
     requestCurrentUser,
     requestNotifications,
     selectors,
     useAppDispatch,
-    useAppSelector
+    useAppSelector,
+    useGetSegueEnvironmentQuery
 } from "../../state";
 import {Route, Router, Switch} from "react-router-dom";
 import {Question} from "../pages/Question";
@@ -87,7 +87,7 @@ export const IsaacApp = () => {
     const dispatch = useAppDispatch();
     const serverError = useAppSelector((state: AppState) => state && state.error && state.error.type == "serverError" || false);
     const goneAwayError = useAppSelector((state: AppState) => state && state.error && state.error.type == "goneAwayError" || false);
-    const {data: segueEnvironment} = isaacApi.endpoints.getSegueEnvironment.useQuery();
+    const {data: segueEnvironment} = useGetSegueEnvironmentQuery();
     const notifications = useAppSelector((state: AppState) => state && state.notifications && state.notifications.notifications || []);
     const user = useAppSelector(selectors.user.orNull);
