+++ src/app/components/pages/Gameboard.tsx
@@ -1,11 +1,10 @@
 import React, {useEffect} from "react";
 import {
-    isaacApi,
     logAction,
     selectors,
     setAssignBoardPath,
     useAppDispatch,
-    useAppSelector
+    useAppSelector, useGetGameboardByIdQuery
 } from "../../state";
 import {Link, withRouter} from "react-router-dom";
 import {Button, Col, Container, ListGroup, ListGroupItem, Row} from "reactstrap";
@@ -145,7 +144,7 @@ export const GameboardViewer = ({gameboard, className}: {gameboard: GameboardDTO
 export const Gameboard = withRouter(({ location }) => {
     const dispatch = useAppDispatch();
     const gameboardId = location.hash ? location.hash.slice(1) : null;
-    const gameboardQuery = isaacApi.endpoints.getGameboardById.useQuery(gameboardId || skipToken);
+    const gameboardQuery = useGetGameboardByIdQuery(gameboardId || skipToken);
     const { data: gameboard } = gameboardQuery;
     const user = useAppSelector(selectors.user.orNull);
