+++ src/app/components/pages/GameboardFilter.tsx
@@ -3,10 +3,11 @@ import {
     AppState,
     extractDataFromQueryResponse,
     fetchConcepts,
-    isaacApi,
     setAssignBoardPath,
     useAppDispatch,
-    useAppSelector
+    useAppSelector,
+    useGenerateTemporaryGameboardMutation,
+    useLazyGetGameboardByIdQuery
 } from "../../state";
 import {TitleAndBreadcrumb} from "../elements/TitleAndBreadcrumb";
 import {Link, RouteComponentProps, useHistory, withRouter} from "react-router-dom";
@@ -458,8 +459,8 @@ export const GameboardFilter = withRouter(({location}: RouteComponentProps) => {
 
     const [gameboard, setGameboard] = useState<GameboardDTO | NOT_FOUND_TYPE | null | undefined>();
     const gameboardIdAnchor = location.hash ? location.hash.slice(1) : null;
-    const [ generateTemporaryGameboard ] = isaacApi.endpoints.generateTemporaryGameboard.useMutation();
-    const [ loadGameboard ] = isaacApi.endpoints.getGameboardById.useLazyQuery();
+    const [ generateTemporaryGameboard ] = useGenerateTemporaryGameboardMutation();
+    const [ loadGameboard ] = useLazyGetGameboardByIdQuery();
 
     useEffect(() => {
         if (isFound(gameboard) && gameboard.id !== gameboardIdAnchor) {
