+++ src/app/components/pages/quizzes/SetQuizzes.tsx
@@ -1,12 +1,12 @@
 import React, {useEffect, useState} from "react";
 import {
-    isaacApi,
     loadQuizAssignments,
     markQuizAsCancelled,
     selectors,
     showQuizSettingModal,
     useAppDispatch,
-    useAppSelector
+    useAppSelector,
+    useGetGroupsQuery
 } from "../../../state";
 import {Link, RouteComponentProps, withRouter} from "react-router-dom";
 import * as RS from "reactstrap";
@@ -88,7 +88,7 @@ const SetQuizzesPageComponent = ({user, location}: SetQuizzesPageProps) => {
     const [pageTitle, setPageTitle] = useState(siteSpecific((activeTab !== MANAGE_QUIZ_TAB.manage ? "Set" : "Manage") + " Tests", "Manage tests"));
 
     // todo: This is so when the quizAssignments selector tries to augment quizzes with group names, it works. Revisit.
-    const { data: groups } = isaacApi.endpoints.getGroups.useQuery(false);
+    const { data: groups } = useGetGroupsQuery(false);
 
     const quizAssignments = useAppSelector(selectors.quizzes.assignments);
